+++ lib/ts/recipe/passwordless/api/implementation.ts
@@ -1,0 +1,5 @@
+import { APIInterface } from "../";
+
+export default function getAPIImplementation(): APIInterface {
+    throw new Error("TODO: needs impl");
+}