+++ lib/ts/recipe/passwordless/recipeImplementation.ts
@@ -1,0 +1,6 @@
+import { RecipeInterface } from "./types";
+import { Querier } from "../../querier";
+
+export default function getRecipeInterface(_: Querier): RecipeInterface {
+    throw new Error("TODO: needs impl");
+}