+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 .
 
+## [unreleased]
+
+-   Adds passwordless recipe
+-   Adds compatibility with FDI 1.11 and CDI 2.10
+
 ## [8.2.0] - 2021-11-15
 
 ### Added